requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
dataclasses==0.6
Django==4.2.7
django-admin==2.0.2
django-cors-headers==4.3.0
django-excel-response2==3.0.6
django-filter==23.3
django-restframework==0.0.1
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.2.31
dnspython==2.4.2
excel-base==1.0.4
firebase-admin==6.2.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
isoweek==1.3.3
msgpack==1.0.7
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rsa==4.9
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
xlwt==1.3.0
